# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/item/upsert": {
            "post": {
                "operationId": "2333afe0bef6fb381a425cdc1cd105ab",
                "parameters": [
                    {
                        "name": "X-Custom-Header",
                        "in": "header",
                        "description": "Custom header for additional information",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "groupid",
                                    "itemname",
                                    "itemprice"
                                ],
                                "properties": {
                                    "itemid": {
                                        "description": "Item ID (when provided, item will be updated, otherwise created)",
                                        "type": "string"
                                    },
                                    "groupid": {
                                        "description": "Group ID",
                                        "type": "string"
                                    },
                                    "itemname": {
                                        "description": "Item Name",
                                        "type": "string"
                                    },
                                    "itemdescription": {
                                        "description": "Item Description",
                                        "type": "string"
                                    },
                                    "itemprice": {
                                        "description": "Item Price",
                                        "type": "number",
                                        "format": "double",
                                        "maximum": 999999999999.99,
                                        "minimum": 0
                                    },
                                    "itemimage": {
                                        "description": "Item Image",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Saved"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Invalid Token (Usually Occurs when token is expired, which requires re-login)"
                    },
                    "404": {
                        "description": "Invalid Group ID <br/>Invalid Item ID"
                    }
                }
            }
        },
        "/api/item/delete": {
            "post": {
                "operationId": "fb6909740a0b8a88d1466433a17110d3",
                "parameters": [
                    {
                        "name": "X-Custom-Header",
                        "in": "header",
                        "description": "Custom header for additional information",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "groupid",
                                    "itemid"
                                ],
                                "properties": {
                                    "groupid": {
                                        "description": "Group ID",
                                        "type": "string"
                                    },
                                    "itemid": {
                                        "description": "Item ID",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Saved"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Invalid Token (Usually Occurs when token is expired, which requires re-login)"
                    },
                    "404": {
                        "description": "Invalid Group ID <br/>Invalid Item"
                    }
                }
            }
        },
...
    }
}